tratamento de exceções nas telas de Limpeza, Enfermeiros e Agendamento

+++ entidade/enfermeiro.py
@@ -2,5 +2,5 @@ from entidade.funcionario import Funcionario
 
 
 class Enfermeiro(Funcionario):
-    def __init__(self, nome: str, cpf: str, idade: int, rua: str, numero: int, complemento: str, matricula: int, salario: float):
+    def __init__(self, nome: str, cpf: str, idade: int, rua: str, numero: int, complemento: str, matricula: int, salario: int):
         super().__init__(nome, cpf, idade, rua, numero, complemento, matricula, salario)
