+++ entidade/funcionario.py
@@ -4,11 +4,11 @@ from entidade.pessoa import Pessoa
 
 class Funcionario(Pessoa, ABC):
     @abstractmethod
-    def __init__(self, nome: str, cpf: str, idade: int, rua: str, numero: int, complemento: str, matricula: int, salario: float):
+    def __init__(self, nome: str, cpf: str, idade: int, rua: str, numero: int, complemento: str, matricula: int, salario: int):
         super().__init__(nome, cpf, idade, rua, numero, complemento)
         if isinstance(matricula, int):
             self.__matricula = matricula
-        if isinstance(salario, float):
+        if isinstance(salario, int):
             self.__salario = salario
 
     @property
@@ -25,6 +25,6 @@ class Funcionario(Pessoa, ABC):
         return self.__salario
 
     @salario.setter
-    def salario(self, salario: float):
-        if isinstance(salario, float):
+    def salario(self, salario: int):
+        if isinstance(salario, int):
             self.__salario = salario
